Isolate bundle RPC implementation and test
